+++ tests/test_kolmogorov.py
@@ -68,6 +68,7 @@ def test_kolmogorov():
     kolm2 = galsim.Kolmogorov(lam_over_r0=1.5, flux=test_flux, gsparams=gsp)
     assert kolm2 != kolm
     assert kolm2 == kolm.withGSParams(gsp)
+    assert kolm2 == kolm.withGSParams(xvalue_accuracy=1.e-8, kvalue_accuracy=1.e-8)
 
     check_basic(kolm, "Kolmogorov")